Start of weight updates
- Changed topic rating from INT to Float for added precision
- Added Comments Section to article details
- Added Delete capability to comment section
- Added Sentiment Analysis to comments
- Added Alerts after actions
- Added Saved Articles Page
- Change unsave Article function to redirect user to which ever page they unsaved from
- Added No Comments yet alert
- Added Function to update topic weights via actions
- Added weight limit of 8
- Added perform action function to log actions

+++ Recommendations/models.py
@@ -11,6 +11,7 @@ class Topic(models.Model):
 
 class TopicPreference(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    rating = models.IntegerField(default=0)
+    rating = models.FloatField(default=0) 
     topic = models.ForeignKey(Topic, on_delete=models.DO_NOTHING, null=True, blank=True, default=None)
+
